metadata/requirements.txt
requests
astropy
astroquery==0.4.6
asyncio
asyncpg
packaging
python-dotenv
